requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
Pillow==10.1.0
moviepy==1.0.3
opencv-python==4.8.1.78